@@ -32,30 +32,30 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^17.3.12",
-    "@angular/cdk": "^17.3.10",
-    "@angular/common": "^17.3.12",
-    "@angular/compiler": "^17.3.12",
-    "@angular/core": "^17.3.12",
-    "@angular/forms": "^17.3.12",
-    "@angular/localize": "^17.3.12",
-    "@angular/material": "^17.3.10",
-    "@angular/platform-browser": "^17.3.12",
-    "@angular/platform-browser-dynamic": "^17.3.12",
-    "@angular/platform-server": "^17.3.12",
-    "@angular/pwa": "^17.3.10",
-    "@angular/router": "^17.3.12",
-    "@angular/service-worker": "^17.3.12",
-    "@angular/ssr": "^17.3.10",
+    "@angular/animations": "^18.2.13",
+    "@angular/cdk": "^18.2.13",
+    "@angular/common": "^18.2.13",
+    "@angular/compiler": "^18.2.13",
+    "@angular/core": "^18.2.13",
+    "@angular/forms": "^18.2.13",
+    "@angular/localize": "^18.2.13",
+    "@angular/material": "^18.2.13",
+    "@angular/platform-browser": "^18.2.13",
+    "@angular/platform-browser-dynamic": "^18.2.13",
+    "@angular/platform-server": "^18.2.13",
+    "@angular/pwa": "^18.2.12",
+    "@angular/router": "^18.2.13",
+    "@angular/service-worker": "^18.2.13",
+    "@angular/ssr": "^18.2.12",
     "@apollo/client": "^3.7.14",
     "@bhayward93/ckeditor5-build-minds": "^1.1.8",
     "@ckeditor/ckeditor5-angular": "^5.2.0",
     "@ctrl/ngx-emoji-mart": "^9.2.0",
-    "@danielmoncada/angular-datetime-picker": "^17.0.0",
+    "@danielmoncada/angular-datetime-picker": "^18.0.0",
     "@fontsource/inter": "^4.5.15",
     "@gorniv/ngx-universal": "^2.2.3",
     "@mindsorg/web3modal-angular": "^17.3.12",
-    "@ng-bootstrap/ng-bootstrap": "^16.0.0",
+    "@ng-bootstrap/ng-bootstrap": "^17.0.1",
     "@popperjs/core": "^2.9.3",
     "@sentry/angular": "^8.26.0",
     "@sentry/node": "^8.26.0",
@@ -66,7 +66,7 @@
     "@types/yargs": "^17.0.2",
     "@walletconnect/web3-provider": "^1.4.1",
     "angular-plotly.js": "^5.2.2",
-    "angularx-qrcode": "^17.0.1",
+    "angularx-qrcode": "^18.0.2",
     "apollo-angular": "^6.0.0",
     "assert": "^2.0.0",
     "blurhash": "^1.1.4",
@@ -93,7 +93,7 @@
     "medium-editor": "^5.23.3",
     "moment": "^2.29.1",
     "ngx-drag-drop": "^17.0.0",
-    "ngx-markdown": "^17.2.1",
+    "ngx-markdown": "^18.1.0",
     "ngx-plyr-mg": "1.0.1",
     "ngx-popperjs": "^17.0.0",
     "node-cache": "^5.1.2",
@@ -105,7 +105,7 @@
     "raw-loader": "^4.0.2",
     "rxjs": "^7.8.0",
     "socket.io-client": "^4.7.5",
-    "storybook-addon-angular-router": "^1.10.0",
+    "storybook-addon-angular-router": "^1.10.1",
     "stream-browserify": "^3.0.0",
     "stream-http": "^3.2.0",
     "textarea-caret": "^3.1.0",
@@ -119,10 +119,10 @@
     "zone.js": "^0.14.10"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^17.3.10",
-    "@angular/cli": "^17.3.10",
-    "@angular/compiler-cli": "^17.3.12",
-    "@angular/language-service": "^17.3.12",
+    "@angular-devkit/build-angular": "^18.2.12",
+    "@angular/cli": "^18.2.12",
+    "@angular/compiler-cli": "^18.2.13",
+    "@angular/language-service": "^18.2.13",
     "@angularclass/hmr": "^3.0.0",
     "@babel/core": "^7.17.8",
     "@compodoc/compodoc": "^1.1.19",
